requirements.txt
llama-index
llama-index-llms-cohere
llama-index-embeddings-cohere
python-dotenv
requests
Wikipedia-API
qdrant_client
llama-index-vector-stores-qdrant
pydantic
